requirements.txt
langgraph>=0.2.33
pydantic>=2.7
pandas>=2.2
pyarrow>=15.0
sqlglot>=23.0
jinja2>=3.1
rich>=13.7
click>=8.1
python-dotenv>=1.0
google-cloud-bigquery>=3.25
google-auth>=2.30
google-generativeai>=0.6
db-dtypes==1.2.0